--- elevenProject.xcodeproj/xcuserdata/luttoli.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Bucket
-   uuid = "B4974E46-11E4-4238-BD28-1829E3372B86"
-   type = "1"
-   version = "2.0">
-</Bucket>
