+++ elevenProject/main.swift
@@ -1,4 +1,3 @@
 import Foundation
 
-//앱 스타트
 appStart()
