+++ elevenProject/walt/Bank/bank.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+func bank() {
+    
+}
